requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.0.200
openai==0.28.0
chromadb==0.3.29
beautifulsoup4==4.12.2
requests==2.31.0
python-multipart==0.0.9
tiktoken==0.5.1
numpy==1.26.2
pandas==2.1.3
pydantic==1.10.13
jinja2==3.1.2
langchainplus-sdk==0.0.20
PyPDF2==3.0.1 